changing data type of account_id to varchar(32) across all tables

+++ sql/tables.sql
@@ -27,7 +27,7 @@ CREATE TABLE orders (
     symbol           VARCHAR(20)  NOT NULL REFERENCES instruments(symbol),
     side             VARCHAR(4) NOT NULL,
     quantity         INT NOT NULL,
-    price            INT NOT NULL,
+    price            NUMERIC(18,2) NOT NULL,
     order_status     VARCHAR(20) NOT NULL,
     idempotency_key  VARCHAR(100) UNIQUE
 );
